Main program controlling Monitor WORKING!
A program that creates a Monitor and all the other required instances,
adds patterns, emails, via the entities classes, starts monitors via the
Peers class, and stops all the monitors again after a certain amount of
time.

The detection of login state, login, starting and stopping of monitors
works correctly.