Adapt to new way of reading messages, NOW WORKING FOR REAL
In the previous commit it was actually not working. Only one of the
monitor processes received the NewMessage updates from Telegram. So it
would have worked for a single monitored peer, but not for multiple
ones.

With the new way of reading messages (a single process which reads out
all the messages, and filtering the wanted messages in the Java code),
it also works correctly with multiple peers.